==> campaign_touch_points/3.5_clean_lead_table.ipynb <==
import janitor
import pandas as pd
import re
from datetime import datetime



leads = pd.read_csv('data/all_time_leads.csv').clean_names()


selected_leads = leads[['lead_id','mobile_primary',
                        'business_phone_line_1','brizo_id', 
                        'created_date',
                        'first_mel_timestamp','latest_mql_timestamp',
                        'assigned_time_ae_',
                        'opportunity_created_date',
                        'closed_won_date',
                        'stage','unqualified_reason', 'closed_lost_reason']].drop_duplicates()


selected_leads['mobile_primary'] = selected_leads['mobile_primary'].astype(str)
selected_leads['business_phone_line_1'] = selected_leads['business_phone_line_1'].astype(str)

selected_leads['mobile_primary'] = selected_leads['mobile_primary'].str.replace(r'\D+', '', regex=True)
selected_leads['business_phone_line_1'] = selected_leads['business_phone_line_1'].str.replace(r'\D+', '', regex=True)



def clean_to_10(s):
    # 1) remove any non-digits
    digits = re.sub(r'\D+', '', str(s))
    # 2) if >10 digits and starts with "1", drop the leading "1"
    if len(digits) > 10 and digits.startswith('1'):
        digits = digits[1:]
    # 3) only keep if exactly 10 digits
    return digits if len(digits) == 10 else None

for col in ['mobile_primary', 'business_phone_line_1']:
    selected_leads[col] = selected_leads[col].apply(clean_to_10)


Change timestamp format into date

# date_cols = [
#     'created_date',
#     'first_mel_timestamp',
#     'latest_mql_timestamp',
#     'assigned_time_ae_',
#     'opportunity_created_date',
#     'closed_won_date'
# ]

# date_cols = selected_leads[date_cols]
# date_cols.to_csv('data/exported_data/date_cols.csv', index=False)

timestamp_col = [
    'first_mel_timestamp',
    'latest_mql_timestamp',
    'assigned_time_ae_',
]

for col in timestamp_col:
    selected_leads[col] = pd.to_datetime(selected_leads[col], errors='coerce').dt.date


# Sort the DataFrame by first_mel_timestamp (newest first)
selected_leads = selected_leads.sort_values(by='first_mel_timestamp', ascending=False)

# Assign row number within each ID group
selected_leads['row_num'] = selected_leads.groupby('brizo_id').cumcount() + 1

# Keep only the top row (like QUALIFY row_number = 1)
top_leads = selected_leads[selected_leads['row_num'] == 1].copy()


# inspect
print(top_leads.head(5))

# save
top_leads.to_csv('data/exported_data/deduped_leads_long.csv', index=False)
