==> campaign_touch_points/4_brizo_matching_leads.ipynb <==
import pandas as pd
import janitor
import numpy as np

leads = pd.read_csv('data/exported_data/deduped_leads_long.csv').clean_names()
brizo = pd.read_csv('data/exported_data/combined.csv').clean_names()

### get the date for campaign list sources
brizo_list = brizo['list_source'].unique().tolist()

df = pd.DataFrame(brizo_list, columns=['list_source'])

df['year_month'] = df['list_source'].str.extract(r'(\d{4}-\d{2})')

df.to_csv('data/exported_data/brizo_list_with_ym.csv', index=False)

# 4) Inspect
print(df.head())


# ensure phone_number is string
brizo['phone_number'] = brizo['phone_number'].astype(str)
# remove any non-digit (spaces, punctuation, letters, etc.)
brizo['phone_number'] = brizo['phone_number'].str.replace(r'\D+', '', regex=True)

brizo_counts = (
    brizo
      .groupby(['brizo_id', 'phone_number'])
      .size()
      .reset_index(name='count')
)
display(brizo_counts)

selected_leads = leads[['lead_id','mobile_primary',
                        'business_phone_line_1','brizo_id', 
                        'created_date',
                        'first_mel_timestamp','latest_mql_timestamp',
                        'assigned_time_ae_',
                        'opportunity_created_date',
                        'closed_won_date',
                        'stage','unqualified_reason', 'closed_lost_reason']].drop_duplicates()



# Count how many times each brizo_id appears
dup_counts = (
    selected_leads
      .groupby('brizo_id')
      .size()
      .reset_index(name='n_matches')
)

# See which IDs appear more than once
print( dup_counts.query('n_matches > 1') )

match_brizo = brizo_counts.merge(
    selected_leads,
    on='brizo_id',
    how='inner'
).assign(match_key='brizo_id')


print(match_brizo.head())


match_brizo.to_csv('data/exported_data/match_brizo.csv', index=False)
